pyproject.toml
[project]
name = "autopilot-expert-multi-agent"
version = "1.0.0"
description = "Multi-agent AI system for Slack with FastAPI backend"
requires-python = ">=3.11"
dependencies = [
    "celery>=5.5.3",
    "fastapi>=0.115.14",
    "google-genai>=1.22.0",
    "kombu>=5.5.4",
    "networkx>=3.5",
    "pinecone-client>=6.0.0",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "pyyaml>=6.0",
    "redis>=6.2.0",
    "slack-sdk>=3.35.0",
    "uvicorn>=0.34.3",
]

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["agents*", "models*", "services*", "tools*", "utils*", "workers*"]
exclude = ["attached_assets*", "__pycache__*", "*.pyc"]

[tool.setuptools.package-data]
"*" = ["*.py"]
